Add support for UDA processing in expression parsing
This commit enhances the expression parsing functionality to also handle UDA (User Defined Attributes). A check is added to discern UDA from regular attributes. If it is a UDA, a different workflow is followed, and if not, the regular path is used. Also, some commented-out debug logging lines were left in the code which can prove useful for future debugging. The feature of let_chains is also enabled in lib.rs.